src/requirements.txt
google-cloud-logging
google-cloud-secret-manager
google-cloud-pubsub
google-auth
google-cloud-language
tweepy
googlemaps